core: imx: caam: remove imx_caam.h header file
Remove imx_caam header file and move its content to source file.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
